Build: Added Ctrl+S save shortcut (overwrited the default behavior)
I built this specifically because, in these kinds of text editing applications, users always auto-press ctrl + s with their muscle memory. Then the browser asks to save the html file. This is so annoying. That's why I replaced the existing behavior and connected it to the builtin save function in the application.